requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep Learning (choose based on your needs)
torch>=2.0.0
tensorflow>=2.13.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking
mlflow>=2.8.0
wandb>=0.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
click>=8.1.0

# Data Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Model Serving
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0